Rename component and update submodules
Renamed to component_data and component_update to avoid ambiguity and keep naming consistent.
