+++ spatialos-sdk/src/worker/schema/component_data.rs
