+++ spatialos-sdk/src/worker/schema/component_update.rs
